requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
yfinance==0.2.18
requests==2.31.0
beautifulsoup4==4.12.2
plotly==5.17.0
openai==1.3.7
python-dotenv==1.0.0
newsapi-python==0.2.6
textblob==0.17.1
scikit-learn==1.3.2
ta==0.10.2
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0 